Create SwayService
This is for the most part a copy-paste job of hyprland.

Uses ToplevelManager to provide information about toplevels which is not
available from the I3 api.

Some features like taskbar focus is a bit broken as this happens by
app_id which falls apart for XWayland windows and applications with
multiple open windows.