fix: Unset `PATH` that is auto-generated by Nix
This addresses a problem where the launcher is
not able to actually launch any applications when
it is started by a systemd unit generated by Nix.

The search path would typically be inherited by a
systemd unit, but Nix's unit generator will
assign a specific, minimal PATH that only includes
the unit's runtime dependencies.